Debugging - feedback read timer shizzle

+++ components/cc1101/fan.h
@@ -27,8 +27,8 @@ class CC1101Fan : public PollingComponent, public fan::Fan {
   void control(const fan::FanCall &call) override;
   void write_state_();
   void publish_state();
-  void resetFanSpeed(uint8_t seconds);
-  void startResetTimer(uint8_t seconds);
+  void resetFanSpeed(uint16_t seconds);
+  void startResetTimer(uint16_t seconds);
   int speed_count_{};
   bool map_off_to_zero_{};
   //fan::FanTraits traits_;
